Populating our options from the API.
