+++ src/views/pages/Contacts/ContactList.jsx
@@ -1,30 +1,5 @@
 import React, {useEffect, useState} from 'react';
-import { BehaviorSubject } from 'rxjs';
-import axios from 'axios';
-
-class ContactsManager {
-  _identities = new BehaviorSubject([]);
-
-  identities$ = this._identities.asObservable();
-
-  async populateContacts() {
-    let results = await axios
-      .get(`http://localhost:1234/api/identities/`, {
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      })
-      .then(function (response) {
-        console.log(response.data.results);
-        return response.data.results;
-      })
-      .catch(function (error) {
-        console.error(error);
-        return [];
-      });
-    this._identities.next([...results]);
-  }
-}
+import ContactsManager from '../../../services/ContactsManager.service';
 
 const contactService = new ContactsManager();
 
@@ -34,7 +9,7 @@ function ContactsList() {
   useEffect(() => {
     const sub = contactService.identities$.subscribe((d) => {
       setContactList(d);
-    })
+    });
 
     contactService.populateContacts();
